@@ -13,13 +13,15 @@
   "author": "Andrea Spadaccini",
   "license": "GPL-2.0-or-later",
   "dependencies": {
-    "@material-ui/core": "^4.12.3",
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
     "@microsoft/applicationinsights-web": "^2.6.1",
+    "@mui/material": "^5.9.0",
     "lodash": "^4.17.19",
     "monaco-editor-webpack-plugin": "^7.0.0",
     "node-forge": "^1.3.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.1.0",
     "react-monaco-editor": "^0.49.0"
   },
